Feat: Implement Boot Completed Receiver for Azan Alarms
Adds a `DeviceBootReceiver` to listen for the `BOOT_COMPLETED` broadcast. This receiver will call `setAllAlarms` to reschedule Azan alarms after the device has restarted.

Also adds a `cancelAllAlarms` extension function to `Context` which is called before setting new alarms to prevent duplicates.

The necessary permission for receiving boot completed broadcasts is added to the manifest.